// dashboard/src/api/types.ts
export interface NodeDTO {
  id: string;
  type: string;
  name: string;
  role_title?: string;
  description?: string;
  parent_id?: string | null;
  enabled?: boolean;
}